01 - Discord Bot/requirements.txt
aiofiles==0.5.0
discord.py==1.4.1
discord==1.0.1
python-dotenv==0.15.0